bookforge/requirements.txt
anthropic>=0.40.0
openai>=1.50.0
google-generativeai>=0.8.0
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0
textstat>=0.7.0
